// src/app/interfaces/caa.ts
export class CAA {
     documentNumber: number;
     answer1: number;
     answer2: number;
     answer3: number;
     answer4: number;
     answer5: number;
     answer6: number;
     answer7: number;
     answer8: number;
     answer9: number;
     answer10: number;
     answer11: number;
     answer12: number;
     answer13: number;
     answer14: number;
     answer15: number;
     answer16: number;
     answer17: number;     
}